Fix: 채팅 리스트 페이지에서 "채팅하기" 버튼 클릭 시 홈으로 이동 후 채팅 생성 상태로 변화
- 원인: useNavigate의 state로 넘겨준 값은 새로고침을 통해 비워지지 않기 때문에 새로고침해도 채팅 생성 상태가 유지되었음
- 해결: 홈에서 나갈 때 채팅 생성 상태의 값을 변경해주던 것을 App.js에서 경로를 확인하고 변경해주는 방법으로 변경하고, 채팅 리스트 페이지에서 "채팅하기" 버튼 클릭 시 홈으로 이동함과 동시에 채팅 생성 상태를 변경시켜줌
