+++ src/App.js
@@ -6,16 +6,26 @@ import ErrorBoundary from "./components/Error/ErrorBoundary";
 import GlobalErrorFallback from "./components/Common/GlobalErrorFallback";
 import { useSetRecoilState } from "recoil";
 import prevChatRoomIdState from "./recoil/atoms/chatRoomId";
+import { createNewChatState, newChatMemberState } from "./recoil/atoms/newChat";
 
 function App() {
   const location = useLocation();
   const setPrevChatRoomId = useSetRecoilState(prevChatRoomIdState);
+  const setNewChatMembers = useSetRecoilState(newChatMemberState);
+  const setCreateNewChat = useSetRecoilState(createNewChatState);
 
   useEffect(() => {
     if (location.pathname !== "/chatList") {
       setPrevChatRoomId(null);
     }
   }, [location, setPrevChatRoomId]);
+
+  useEffect(() => {
+    if (location.pathname !== "/") {
+      setCreateNewChat(false);
+      setNewChatMembers([]);
+    }
+  }, [location, setCreateNewChat, setNewChatMembers]);
   return (
     <Suspense fallback={<Spinner />}>
       <QueryErrorResetBoundary>
