+++ src/recoil/atoms/newChat.js
@@ -1,9 +1,12 @@
 import { atom } from "recoil";
 import { NEW_CHAT_KEYS } from "../../constants/recoilKeys";
 
-const createNewChatState = atom({
+export const createNewChatState = atom({
   key: NEW_CHAT_KEYS.createNewChat,
   default: false,
 });
 
-export default createNewChatState;
+export const newChatMemberState = atom({
+  key: NEW_CHAT_KEYS.newChatMembers,
+  default: [],
+});
