--- src/recoil/atoms/newChatMember.js
@@ -1,9 +1,0 @@
-import { atom } from "recoil";
-import { NEW_CHAT_KEYS } from "../../constants/recoilKeys";
-
-const newChatMemberState = atom({
-  key: NEW_CHAT_KEYS.newChatMembers,
-  default: [],
-});
-
-export default newChatMemberState;
